move clients to pages
