+++ src/pages/Businesses/index.js
@@ -1,0 +1,2 @@
+export { default as BusinessesDash } from './BusinessesDash';
+export { default as BusinessesGen } from './BusinessesGen';
