+++ src/pages/Clients/ClientsGen.jsx
@@ -2,12 +2,12 @@ import React, { useState, useEffect } from 'react';
 import { connect } from 'react-redux';
 import { useHistory, useParams } from 'react-router-dom';
 import { Button } from '@material-ui/core';
-import { SaveAlert } from '../Common/Alerts';
 import { convertKeysCase } from '../../utils/caseConversion';
 import {
     createClientAct,
     updateClientByIdAct,
 } from '../../state/clients/clientActions';
+import { SaveAlert } from '../../components/Common/Alerts';
 import ClientsTemplate from './ClientsTemplate';
 
 const InitialForm = {
