+++ src/pages/Clients/index.js
@@ -1,3 +1,2 @@
 export { default as ClientsDash } from './ClientsDash';
 export { default as ClientsGen } from './ClientsGen';
-export { default as ClientsTemplate } from './ClientsTemplate';
